Move GetSettleBalance to tomox_state package to reuse in order validation
